Break up suffix varation into three different ones
