+++ src/main/java/org/immregistries/smm/transform/procedure/SuffixVariationNumbers.java
@@ -1,0 +1,11 @@
+package org.immregistries.smm.transform.procedure;
+
+public class SuffixVariationNumbers extends AbstractSuffixVariation {
+
+  protected static final String[] SUFFIXES = new String[] {"2nd", "3rd"};
+  
+  @Override
+  protected String[] getSuffixes() {
+    return SUFFIXES;
+  }
+}
