+++ src/main/java/org/immregistries/smm/transform/procedure/SuffixVariationPeriods.java
@@ -1,0 +1,11 @@
+package org.immregistries.smm.transform.procedure;
+
+public class SuffixVariationPeriods extends AbstractSuffixVariation {
+
+  protected static final String[] SUFFIXES = new String[] {"Jr.", "Sr."};
+  
+  @Override
+  protected String[] getSuffixes() {
+    return SUFFIXES;
+  }
+}
